fix check-declare error by eliminating declare-function
The declare-function statement was probably intended to silence the
byte compiler about the use of 'migemo-forward' from the optional
'migemo' package.  However, the syntax of the declare-function was
invalid, so rather than solving the byte compile error this code
created a check-declare error as well.

This change drops the declare-function line from the file. Instead, we
wrap 'with-no-warnings' around the call to 'migemo-foward'. This is
sufficient to silence the byte-compiler. At runtime, the predicate
'anzu--use-migemo-p' ensures this line is never called unless the
'migemo' package is actually present.